requirements.txt
uvicorn==0.34.0
fastapi==0.115.8
sqlalchemy==2.0.38
pymysql==1.1.1
python-dotenv==1.0.1
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
